+++ substratools/__init__.py
@@ -4,6 +4,7 @@ from . import function
 from . import opener
 from .function import execute
 from .function import load_performance
+from .function import register
 from .function import save_performance
 from .opener import Opener
 
@@ -14,5 +15,6 @@ __all__ = [
     Opener,
     execute,
     load_performance,
+    register,
     save_performance,
 ]