+++ tests/test_function.py
@@ -25,6 +25,7 @@ def setup(valid_opener):
     pass
 
 
+@function.register
 def train(
     inputs: TypedDict(
         "inputs",
@@ -59,6 +60,7 @@ def train(
     utils.save_model(model=new_model, path=outputs.get(OutputIdentifiers.model))
 
 
+@function.register
 def predict(
     inputs: TypedDict("inputs", {InputIdentifiers.datasamples: Any, InputIdentifiers.model: List[PathLike]}),
     outputs: TypedDict("outputs", {OutputIdentifiers.predictions: PathLike}),
@@ -77,6 +79,7 @@ def predict(
     utils.save_predictions(predictions=pred, path=outputs.get(OutputIdentifiers.predictions))
 
 
+@function.register
 def no_saved_train(inputs, outputs, task_properties):
     # TODO: checks on data
     # load models
@@ -97,6 +100,7 @@ def no_saved_train(inputs, outputs, task_properties):
     utils.no_save_model(model=new_model, path=outputs.get(OutputIdentifiers.model))
 
 
+@function.register
 def wrong_saved_train(inputs, outputs, task_properties):
     # TODO: checks on data
     # load models
@@ -217,16 +221,15 @@ def test_execute_train(workdir, output_model_path):
 
     assert not output_model_path.exists()
 
-    function.execute(train, sysargs=["--function-name", "train"] + options)
+    function.execute(sysargs=["--function-name", "train"] + options)
     assert output_model_path.exists()
 
     function.execute(
-        train,
-        sysargs=["--function-name", "train", "--fake-data", "--n-fake-samples", "1", "--outputs", json.dumps(outputs)],
+        sysargs=["--function-name", "train", "--fake-data", "--n-fake-samples", "1", "--outputs", json.dumps(outputs)]
     )
     assert output_model_path.exists()
 
-    function.execute(train, sysargs=["--function-name", "train", "--log-level", "debug"] + options)
+    function.execute(sysargs=["--function-name", "train", "--log-level", "debug"] + options)
     assert output_model_path.exists()
 
 
@@ -251,7 +254,7 @@ def test_execute_train_multiple_models(workdir, output_model_path, create_models
     command = ["--function-name", "train"]
     command.extend(options)
 
-    function.execute(train, sysargs=command)
+    function.execute(sysargs=command)
     assert output_model_path.exists()
     with open(output_model_path, "r") as f:
         model = json.load(f)
@@ -275,7 +278,7 @@ def test_execute_predict(workdir, output_model_path, create_models, valid_opener
     assert not pred_path.exists()
     command = ["--function-name", "train"]
     command.extend(train_options)
-    function.execute(train, predict, sysargs=command)
+    function.execute(sysargs=command)
     assert output_model_path.exists()
 
     # do predict on output model
@@ -287,7 +290,7 @@ def test_execute_predict(workdir, output_model_path, create_models, valid_opener
     pred_options = ["--inputs", json.dumps(pred_inputs), "--outputs", json.dumps(pred_outputs)]
 
     assert not pred_path.exists()
-    function.execute(train, predict, sysargs=["--function-name", "predict"] + pred_options)
+    function.execute(sysargs=["--function-name", "predict"] + pred_options)
     assert pred_path.exists()
     with open(pred_path, "r") as f:
         pred = json.load(f)
@@ -308,11 +311,7 @@ def test_execute_predict(workdir, output_model_path, create_models, valid_opener
     pred_options = ["--inputs", json.dumps(pred_inputs), "--outputs", json.dumps(pred_outputs)]
 
     assert not pred_path.exists()
-    function.execute(
-        train,
-        predict,
-        sysargs=["--function-name", "predict"] + pred_options,
-    )
+    function.execute(sysargs=["--function-name", "predict"] + pred_options)
     assert pred_path.exists()
     with open(pred_path, "r") as f:
         pred = json.load(f)
@@ -328,11 +327,17 @@ def test_model_check(valid_function_workspace, function_to_run):
 
 
 def test_function_not_found():
-    def train():
+    with pytest.raises(exceptions.FunctionNotFoundError):
+        function.execute(sysargs=["--function-name", "imaginary_function"])
+
+
+def test_function_name_already_register():
+    @function.register
+    def fake_function():
         pass
 
-    with pytest.raises(exceptions.FunctionNotFoundError):
-        function.execute(
-            train,
-            sysargs=["--function-name", "imaginary_function"],
-        )
+    with pytest.raises(exceptions.ExistingRegisteredFunctionError):
+
+        @function.register
+        def fake_function():
+            pass